it feels a bit unstable, but better than it was

+++ core/hardware/driver.h
@@ -36,7 +36,7 @@ struct DRIVER
 
 typedef struct 
 {
-    unsigned short command;
+    unsigned int command;
     unsigned int parameter1;
     unsigned int parameter2;
     unsigned int parameter3;
